+++ example-src/components/FormatExample.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script>
-import TabCompleteInput from '../../../dist/tab-complete-input'
+import TabCompleteInput from '../tab-complete-input'
 import { getFormat, staticList } from './shared'
 
 export default {
