+++ example-src/components/shared.js
