+++ example-src/shims-vue.d.ts
